Move the issue validator's test data to a subfolder
