+++ tests/__data__/input/issue/validate/data/categories.csv
