+++ tests/__data__/input/issue/validate/data/channels.csv
