+++ tests/__data__/input/issue/validate/data/logos.csv
