+++ tests/__data__/input/issue/validate/issues.js
